Begin chapter 04 - Interesting map
For now most of the code in `main.rs` was split in proper modules. The
biggest difference is that most of the components, functions and
properties are qualified as `pub`. Other than that, it is the exact
same code.

One thing that I learned here was that you can use the `main` module's
function by importing `super`.